--- src/components/Request/RequestModal.tsx
@@ -1,268 +1,0 @@
-import { Box, useTheme, Typography } from '@mui/material';
-import { useState, useEffect } from 'react';
-import AppFormModal from '../common/AppFormModal';
-import AppInputField from '../common/AppInputField';
-import AppDropdown from '../common/AppDropdown';
-import BasicDatePicker from '../common/BasicDatePicker';
-
-interface Request {
-  id: number;
-  title: string;
-  type: string;
-  status: string;
-  startDate: string;
-  endDate: string;
-  reason: string;
-}
-
-function RequestModal({
-  open,
-  onClose,
-  title,
-  initialData,
-}: {
-  open: boolean;
-  onClose: () => void;
-  title: string;
-  initialData?: Request | null;
-}) {
-  const theme = useTheme();
-
-  const controlBg = theme.palette.background.paper;
-  const direction = theme.direction;
-
-  // State variables
-  const [titleVal, setTitleVal] = useState('');
-  const [reqType, setReqType] = useState('wfh');
-  const [fromDate, setFromDate] = useState<string | null>(null);
-  const [toDate, setToDate] = useState<string | null>(null);
-  const [reason, setReason] = useState('');
-  const [additionalDetails, setAdditionalDetails] = useState('');
-
-  // Sync state when initialData or open status changes
-  useEffect(() => {
-    if (open) {
-      if (initialData) {
-        setTitleVal(initialData.title || '');
-        setReqType(
-          initialData.type?.toLowerCase() === 'wfh' ||
-            initialData.type?.toLowerCase() === 'work from home'
-            ? 'wfh'
-            : 'leave'
-        );
-        setFromDate(initialData.startDate);
-        setToDate(initialData.endDate);
-
-        // Try to match reason if it's one of the keys, otherwise leave empty or handle mapping
-        const r = initialData.reason?.toLowerCase() || '';
-        if (['personal', 'official', 'other'].includes(r)) {
-          setReason(r);
-        } else {
-          setReason('other');
-          setAdditionalDetails(initialData.reason || '');
-        }
-      } else {
-        // Reset for new request
-        setTitleVal('');
-        setReqType('wfh');
-        setFromDate(null);
-        setToDate(null);
-        setReason('');
-        setAdditionalDetails('');
-      }
-    }
-  }, [initialData, open]);
-
-  const getLabel = (en: string, ar: string) => (direction === 'rtl' ? ar : en);
-
-  const handleSubmit = () => {
-    // Handle form submission logic here
-    console.log({
-      id: initialData?.id,
-      titleVal,
-      reqType,
-      fromDate,
-      toDate,
-      reason,
-      additionalDetails,
-    });
-    onClose();
-  };
-
-  return (
-    <AppFormModal
-      title={title}
-      open={open}
-      onClose={onClose}
-      onSubmit={handleSubmit}
-      submitLabel={
-        initialData
-          ? getLabel('Save Changes', 'حفظ التغييرات')
-          : getLabel('Submit Request', 'إرسال الطلب')
-      }
-      cancelLabel={getLabel('Cancel', 'إلغاء')}
-    >
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          gap: 2,
-          width: '100%',
-        }}
-      >
-        <AppInputField
-          label={getLabel('Title', 'العنوان')}
-          labelClassName='label'
-          value={titleVal}
-          onChange={val =>
-            setTitleVal(typeof val === 'string' ? val : val.target.value)
-          }
-          placeholder={getLabel('Title', 'العنوان')}
-        />
-
-        <AppDropdown
-          label={getLabel('Request Type', 'نوع الطلب')}
-          showLabel={true}
-          labelClassName='label'
-          placeholder={getLabel('Select', 'اختر النوع')}
-          inputBackgroundColor={controlBg}
-          value={reqType}
-          onChange={e => setReqType(String(e.target.value))}
-          options={[
-            {
-              value: 'wfh',
-              label: getLabel('Work From Home', 'العمل من المنزل'),
-            },
-            { value: 'leave', label: getLabel('Leave', 'إجازة') },
-          ]}
-          sx={{
-            '& .MuiOutlinedInput-root': {
-              direction: direction === 'rtl' ? 'rtl' : 'ltr',
-            },
-            '& .MuiSelect-select, & .MuiInputBase-input': {
-              fontSize: 'var(--body-font-size)',
-              lineHeight: 'var(--body-line-height)',
-              letterSpacing: 'var(--body-letter-spacing)',
-              fontWeight: 400,
-            },
-          }}
-        />
-
-        <Box
-          sx={{
-            display: 'flex',
-            gap: 2,
-            flexDirection: { xs: 'column', sm: 'row' },
-          }}
-        >
-          <BasicDatePicker
-            label={getLabel('From Date', 'تاريخ البدء')}
-            value={fromDate}
-            labelClassName='label'
-            onChange={date => setFromDate(date)}
-            placeholder={getLabel('Select date', 'اختر التاريخ')}
-            required
-          />
-          <BasicDatePicker
-            label={getLabel('To Date', 'تاريخ الانتهاء')}
-            value={toDate}
-            labelClassName='label'
-            onChange={date => setToDate(date)}
-            placeholder={getLabel('Select date', 'اختر التاريخ')}
-            required
-          />
-        </Box>
-
-        <AppDropdown
-          label={getLabel('Reason', 'السبب')}
-          showLabel={true}
-          labelClassName='label'
-          placeholder={getLabel('Select', 'اختر السبب')}
-          inputBackgroundColor={controlBg}
-          value={reason}
-          onChange={e => setReason(String(e.target.value))}
-          options={[
-            { value: 'personal', label: getLabel('Personal', 'شخصي') },
-            { value: 'official', label: getLabel('Official', 'رسمي') },
-            { value: 'other', label: getLabel('Other', 'أخرى') },
-          ]}
-          sx={{
-            '& .MuiOutlinedInput-root': {
-              direction: direction === 'rtl' ? 'rtl' : 'ltr',
-            },
-            '& .MuiSelect-select, & .MuiInputBase-input': {
-              fontSize: 'var(--body-font-size)',
-              lineHeight: 'var(--body-line-height)',
-              letterSpacing: 'var(--body-letter-spacing)',
-              fontWeight: 400,
-            },
-          }}
-        />
-
-        <AppInputField
-          label={getLabel(
-            'Additional Details (Optional)',
-            'تفاصيل إضافية (اختياري)'
-          )}
-          labelClassName='label'
-          value={additionalDetails}
-          onChange={val =>
-            setAdditionalDetails(
-              typeof val === 'string' ? val : val.target.value
-            )
-          }
-          placeholder={getLabel(
-            'Enter any additional details...',
-            'أدخل أي تفاصيل إضافية...'
-          )}
-        />
-
-        {/* Conditional Info Box for WFH */}
-        {reqType === 'wfh' && (
-          <Box
-            sx={{
-              backgroundColor:
-                theme.palette.mode === 'dark'
-                  ? 'var(--black-color)'
-                  : 'var(--primary-color)',
-              borderLeft: `4px solid ${theme.palette.primary.main}`,
-              borderRadius: 'var(--border-radius-2xl)',
-              padding: '12px 16px',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 0.5,
-              mt: 1,
-            }}
-          >
-            <Typography
-              sx={{
-                fontSize: '14px',
-                fontWeight: 600,
-                color: theme.palette.text.primary,
-              }}
-            >
-              {getLabel(
-                'Your availability during WFH',
-                'تواجدك خلال العمل من المنزل'
-              )}
-            </Typography>
-            <Typography
-              sx={{
-                fontSize: '13px',
-                color: theme.palette.text.primary,
-                lineHeight: '1.4',
-              }}
-            >
-              {getLabel(
-                "You'll remain available on Slack, email, and video calls as per office hours (9 AM - 6 PM).",
-                'ستبقى متاحاً على Slack والبريد الإلكتروني ومكالمات الفيديو حسب ساعات العمل (9 صباحاً - 6 مساءً).'
-              )}
-            </Typography>
-          </Box>
-        )}
-      </Box>
-    </AppFormModal>
-  );
-}
-
-export default RequestModal;
