+++ src/components/common/AppDropdown.tsx
@@ -112,8 +112,8 @@ const AppDropdown = React.forwardRef<HTMLDivElement, AppDropdownProps>(
               }
               className={labelClassName}
               sx={{
-                // fontWeight: 500,
-                // fontSize: 'var(--subheading3-font-size)',
+                fontWeight: 500,
+                fontSize: 'var(--subheading3-font-size)',
                 lineHeight: 'var(--subheading2-line-height)',
                 letterSpacing: 'var(--subheading2-letter-spacing)',
                 color:
